Sync automation to release-4.3 (#4514)
Add Sync automation for tyk

:robot: Beep, boop, this PR has been generated by
[automation](https://github.com/TykTechnologies/gromit/blob/master/cmd/policy.go)
on Thu Dec 8 21:10:59 UTC 2022
because commits on `master` need to be propagated to the branches
`[release-4 release-4.3]`
as per the current repository policy.

## Policy for tyk
Commits affecting these files will be ported as listed below.


### Ports
| Source branch | Destination branches |
|---------------|----------------------|
master |  [release-4 release-4.3] |

Thank you for reading this far. Since this PR is generated by
automation, if there
are inconsistencies in the information provided above, they are bugs and
merging
this PR will probably do some damage. Please log a jira in the TD
project citing this PR.

Co-authored-by: Gromit <policy@gromit>