[TD-1474]: Sync releng templates. (#4821)
This is generated from the WIP branch
`fix/td-1528/update-releng-templates` on gromit.
The gromit branch/PR should be merged to master before this PR gets
merged,as sync automation would run on tyk once this is merged.

For all the versioned releng templates to work with sync-automation -
the sync automation script requires the PR that changed the templates be
merged to gromit master.